Projects/Basic ChatBot using Azure OpenAI/requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
streamlit